// src/interfaces/IClients.ts
export interface IClient{
  id: string
  firstDate: string
  lastDate: string
  name: string
  accessories: string
  approval: string
  done: string
  brand: string
  budgetComponent: string
  budgetLabel: string
  changedEquipments: string
  city: string
  defect: string
  district: string
  equipment: string
  fone: string
  local: string
  model: string
  paid: string
  technicalEvaluationt: string
}